requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
pandas==1.5.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.0
sqlalchemy==1.4.46
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0

# 🟨 Data Science & Machine Learning Libraries
matplotlib==3.10.0
numpy==1.26.0
scipy==1.14.0
scikit-learn==1.5.0
seaborn==0.13.0
statsmodels==0.14.0

# 📊 Advanced Analytics & ML
prince==0.10.0
